line related methods in line.js
